GNSS change only: remove need to call uGnssPosGetStop() after uGnssPosGetStart().
Though unstated, it was necessary to call uGnssPosGetStop() before uGnssPosGetStart() could be called once more.  This is now fixed, so uGnssPosGetStart() can be called at any time and the new request will replace the first, the description of uGnssPosGetStart() has been updated and the gnssPosPos() test updated to ensure it remains true.