GNSS fix for uGnssPosStreamedStart() when using M10 over UART.
The setting of the message rate, which is needed by uGnssPosStreamedStart(), is port-specific (i.e. each message rate is different for I2C v SPI v UART1 v UART2).  We test it over I2C (on M8, M9 and M10) and SPI (on M8 and M9, since M10 doesn't have SPI), however we _don't_ test it over UART and we weren't incrementing the value correctly for the UART case.